+++ src-core/src/os/sounds_gen.rs
@@ -1,0 +1,12 @@
+/* THIS FILE IS GENERATED. DO NOT EDIT IT MANUALLY. */
+pub static SOUND_FILES: &[&str] = &[
+    "material_alarm_gentle",
+    "material_alarm_gentle_short_1",
+    "material_alarm_gentle_short_2",
+    "material_hero_simple-celebration-01",
+    "mic_mute",
+    "mic_unmute",
+    "notification_bell",
+    "notification_block",
+    "notification_reverie",
+];