+++ src-ui/app/views/dashboard-view/views/join-notifications-view/join-notifications-view.component.ts
@@ -147,7 +147,10 @@ export class JoinNotificationsViewComponent implements OnInit {
   }
 
   async testSound() {
-    await this.notifications.playSound('notification_reverie', this.config.joinSoundVolume / 100);
+    await this.notifications.playSound(
+      'material_alarm_gentle_short_1',
+      this.config.joinSoundVolume / 100
+    );
     this.playingTestSound = true;
     if (this.playingTestSoundTimeout) clearTimeout(this.playingTestSoundTimeout);
     this.playingTestSoundTimeout = setTimeout(() => {