+++ src-ui/app/views/dashboard-view/views/nightmare-detection-view/nightmare-detection-view.component.ts
@@ -12,6 +12,7 @@ import { map } from 'rxjs';
 import { clamp } from '../../../../utils/number-utils';
 import { hshrink } from '../../../../utils/animations';
 import { NotificationService } from '../../../../services/notification.service';
+import { NIGHTMARE_DETECTION_NOTIFICATION_SOUND } from '../../../../services/nightmare-detection-automation.service';
 
 @Component({
   selector: 'app-nightmare-detection-view',
@@ -68,13 +69,16 @@ export class NightmareDetectionViewComponent implements OnInit {
   }
 
   async testSound() {
-    await this.notifications.playSound('notification_reverie', this.config.soundVolume / 100);
+    await this.notifications.playSound(
+      NIGHTMARE_DETECTION_NOTIFICATION_SOUND,
+      this.config.soundVolume / 100
+    );
     this.playingTestSound = true;
     if (this.playingTestSoundTimeout) clearTimeout(this.playingTestSoundTimeout);
     this.playingTestSoundTimeout = setTimeout(() => {
       this.playingTestSound = false;
       this.playingTestSoundTimeout = undefined;
-    }, 8000);
+    }, 21000);
   }
 
   async onChangeDuration(value: number, unit?: SelectBoxItem) {